Minor update - The script did not tell the user how to end the program
The program will end if the user enters q, quit, or stop. This information is now displayed.

I also found that if a user does not enter q and all the images are shown and closed, the program will remain open until the user enters q or some other stop command. It is getting late so this is the best I can do for now.